# src/main/java/sk/ardevop/nlp/skquadmanager/entity/Dataset.java
package sk.ardevop.nlp.skquadmanager.entity;

import javax.persistence.CascadeType;
import javax.persistence.Entity;
import javax.persistence.FetchType;
import javax.persistence.GeneratedValue;
import javax.persistence.Id;
import javax.persistence.ManyToOne;
import javax.persistence.OneToMany;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;

import java.util.ArrayList;
import java.util.List;
import lombok.ToString;
import org.hibernate.annotations.GenericGenerator;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
@Entity
@ToString(exclude = "corpus")
@EqualsAndHashCode(exclude = "corpus")
public class Dataset {

  @Id
  @GeneratedValue(generator="system-uuid")
  @GenericGenerator(name="system-uuid", strategy = "uuid")
  private String id;
  private String title;
  private String source;

  @ManyToOne(fetch = FetchType.LAZY)
  private Corpus corpus;
  @OneToMany(cascade = CascadeType.REMOVE)
  private List<Paragraph> paragraphs;
}