# src/main/java/sk/ardevop/nlp/skquadmanager/entity/Paragraph.java
package sk.ardevop.nlp.skquadmanager.entity;

import javax.persistence.CascadeType;
import javax.persistence.Entity;
import javax.persistence.FetchType;
import javax.persistence.GeneratedValue;
import javax.persistence.Id;
import javax.persistence.Lob;
import javax.persistence.ManyToOne;
import javax.persistence.OneToMany;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;

import java.util.ArrayList;
import java.util.List;
import lombok.ToString;
import org.hibernate.annotations.GenericGenerator;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
@Entity
@EqualsAndHashCode(exclude = "dataset")
@ToString(exclude = "dataset")
public class Paragraph {

  @Id
  @GeneratedValue(generator="system-uuid")
  @GenericGenerator(name="system-uuid", strategy = "uuid")
  private String id;
  @OneToMany(mappedBy = "paragraph", cascade = CascadeType.REMOVE)
  private List<Question> qas;
  @ManyToOne(fetch = FetchType.LAZY)
  private Dataset dataset;
  @Lob
  private String context;
}